\docType{package}
\name{dashHtmlComponents}
\alias{dashHtmlComponents}
\alias{dashHtmlComponents-package}
\title{dashHtmlComponents: Vanilla HTML Components for Dash}
\description{
For more information:
\url{https://github.com/plotly/dash-html-components}

For issues and bug reports:
\url{https://github.com/plotly/dash-html-components/issues}

For additional help, visit our awesome community forum:
\url{https://community.plot.ly/c/dash}
}
