% Auto-generated: do not edit by hand
\name{htmlForm}

\alias{htmlForm}

\title{Form component}

\description{

}

\usage{
htmlForm(contentEditable=NULL, accept=NULL, autoComplete=NULL, draggable=NULL, tabIndex=NULL, n_clicks_timestamp=NULL, style=NULL, title=NULL, accessKey=NULL, id=NULL, role=NULL, hidden=NULL, method=NULL, spellCheck=NULL, contextMenu=NULL, key=NULL, n_clicks=NULL, target=NULL, lang=NULL, name=NULL, className=NULL, noValidate=NULL, action=NULL, acceptCharset=NULL, dir=NULL, encType=NULL, ...)
}

\arguments{
\item{contentEditable}{Indicates whether the element's content is editable.}

\item{accept}{List of types the server accepts, typically a file type.}

\item{autoComplete}{Indicates whether controls in this form can by default have their values automatically completed by the browser.}

\item{draggable}{Defines whether the element can be dragged.}

\item{tabIndex}{Overrides the browser's default tab order and follows the one specified instead.}

\item{n_clicks_timestamp}{An integer that represents the time (in ms since 1970)
at which n_clicks changed. This can be used to tell
which button was changed most recently.}

\item{style}{Defines CSS styles which will override styles previously set.}

\item{title}{Text to be displayed in a tooltip when hovering over the element.}

\item{accessKey}{Defines a keyboard shortcut to activate or add focus to the element.}

\item{id}{The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{role}{The ARIA role attribute}

\item{hidden}{Prevents rendering of given element, while keeping child elements, e.g. script elements, active.}

\item{method}{Defines which HTTP method to use when submitting the form. Can be GET (default) or POST.}

\item{spellCheck}{Indicates whether spell checking is allowed for the element.}

\item{contextMenu}{Defines the ID of a <menu> element which will serve as the element's context menu.}

\item{key}{A unique identifier for the component, used to improve
performance by React.js while rendering components
See https://reactjs.org/docs/lists-and-keys.html for more info}

\item{n_clicks}{An integer that represents the number of times
that this element has been clicked on.}

\item{target}{}

\item{lang}{Defines the language used in the element.}

\item{name}{Name of the element. For example used by the server to identify the fields in form submits.}

\item{className}{Often used with CSS to style elements with common properties.}

\item{noValidate}{This attribute indicates that the form shouldn't be validated when submitted.}

\item{action}{The URI of a program that processes the information submitted via the form.}

\item{acceptCharset}{List of supported charsets.}

\item{dir}{Defines the text direction. Allowed values are ltr (Left-To-Right) or rtl (Right-To-Left)}

\item{encType}{Defines the content type of the form date when the method is POST.}

\item{...}{wildcards of the form: `data-*` or `aria-*`}
}
