% Auto-generated: do not edit by hand
\name{htmlLink}

\alias{htmlLink}

\title{Link component}

\description{

}

\usage{
htmlLink(contentEditable=NULL, href=NULL, draggable=NULL, integrity=NULL, tabIndex=NULL, n_clicks_timestamp=NULL, style=NULL, crossOrigin=NULL, title=NULL, media=NULL, accessKey=NULL, id=NULL, role=NULL, rel=NULL, hidden=NULL, spellCheck=NULL, contextMenu=NULL, key=NULL, n_clicks=NULL, lang=NULL, sizes=NULL, hrefLang=NULL, className=NULL, dir=NULL, ...)
}

\arguments{
\item{contentEditable}{Indicates whether the element's content is editable.}

\item{href}{The URL of a linked resource.}

\item{draggable}{Defines whether the element can be dragged.}

\item{integrity}{Security Feature that allows browsers to verify what they fetch.     MDN Link}

\item{tabIndex}{Overrides the browser's default tab order and follows the one specified instead.}

\item{n_clicks_timestamp}{An integer that represents the time (in ms since 1970)
at which n_clicks changed. This can be used to tell
which button was changed most recently.}

\item{style}{Defines CSS styles which will override styles previously set.}

\item{crossOrigin}{How the element handles cross-origin requests}

\item{title}{Text to be displayed in a tooltip when hovering over the element.}

\item{media}{Specifies a hint of the media for which the linked resource was designed.}

\item{accessKey}{Defines a keyboard shortcut to activate or add focus to the element.}

\item{id}{The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{role}{The ARIA role attribute}

\item{rel}{Specifies the relationship of the target object to the link object.}

\item{hidden}{Prevents rendering of given element, while keeping child elements, e.g. script elements, active.}

\item{spellCheck}{Indicates whether spell checking is allowed for the element.}

\item{contextMenu}{Defines the ID of a <menu> element which will serve as the element's context menu.}

\item{key}{A unique identifier for the component, used to improve
performance by React.js while rendering components
See https://reactjs.org/docs/lists-and-keys.html for more info}

\item{n_clicks}{An integer that represents the number of times
that this element has been clicked on.}

\item{lang}{Defines the language used in the element.}

\item{sizes}{}

\item{hrefLang}{Specifies the language of the linked resource.}

\item{className}{Often used with CSS to style elements with common properties.}

\item{dir}{Defines the text direction. Allowed values are ltr (Left-To-Right) or rtl (Right-To-Left)}

\item{...}{wildcards of the form: `data-*` or `aria-*`}
}
