% Auto-generated: do not edit by hand
\name{htmlScript}
\alias{htmlScript}
\title{Script component}
\description{

}
\usage{
htmlScript(..., defer=NULL, contentEditable=NULL, draggable=NULL, integrity=NULL, children=NULL, tabIndex=NULL, n_clicks_timestamp=NULL, style=NULL, crossOrigin=NULL, title=NULL, accessKey=NULL, id=NULL, role=NULL, hidden=NULL, type=NULL, spellCheck=NULL, contextMenu=NULL, charSet=NULL, key=NULL, n_clicks=NULL, lang=NULL, src=NULL, className=NULL, async=NULL, dir=NULL)
}
\arguments{
\item{defer}{Indicates that the script should be executed after the page has been parsed.}

\item{contentEditable}{Indicates whether the element's content is editable.}

\item{draggable}{Defines whether the element can be dragged.}

\item{integrity}{Security Feature that allows browsers to verify what they fetch.     MDN Link}

\item{children}{The children of this component}

\item{tabIndex}{Overrides the browser's default tab order and follows the one specified instead.}

\item{n_clicks_timestamp}{An integer that represents the time (in ms since 1970)
at which n_clicks changed. This can be used to tell
which button was changed most recently.}

\item{style}{Defines CSS styles which will override styles previously set.}

\item{crossOrigin}{How the element handles cross-origin requests}

\item{title}{Text to be displayed in a tooltip when hovering over the element.}

\item{accessKey}{Defines a keyboard shortcut to activate or add focus to the element.}

\item{id}{The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{role}{The ARIA role attribute}

\item{hidden}{Prevents rendering of given element, while keeping child elements, e.g. script elements, active.}

\item{type}{Defines the type of the element.}

\item{spellCheck}{Indicates whether spell checking is allowed for the element.}

\item{contextMenu}{Defines the ID of a <menu> element which will serve as the element's context menu.}

\item{charSet}{Declares the character encoding of the page or script.}

\item{key}{A unique identifier for the component, used to improve
performance by React.js while rendering components
See https://reactjs.org/docs/lists-and-keys.html for more info}

\item{n_clicks}{An integer that represents the number of times
that this element has been clicked on.}

\item{lang}{Defines the language used in the element.}

\item{src}{The URL of the embeddable content.}

\item{className}{Often used with CSS to style elements with common properties.}

\item{async}{Indicates that the script should be executed asynchronously.}

\item{dir}{Defines the text direction. Allowed values are ltr (Left-To-Right) or rtl (Right-To-Left)}
}
    