% Auto-generated: do not edit by hand
\name{htmlTextarea}

\alias{htmlTextarea}

\title{Textarea component}

\description{

}

\usage{
htmlTextarea(contentEditable=NULL, n_clicks=NULL, cols=NULL, disabled=NULL, draggable=NULL, wrap=NULL, tabIndex=NULL, n_clicks_timestamp=NULL, style=NULL, rows=NULL, title=NULL, accessKey=NULL, id=NULL, role=NULL, hidden=NULL, spellCheck=NULL, form=NULL, contextMenu=NULL, minLength=NULL, readOnly=NULL, key=NULL, maxLength=NULL, autoFocus=NULL, placeholder=NULL, lang=NULL, autoComplete=NULL, name=NULL, required=NULL, className=NULL, dir=NULL, ...)
}

\arguments{
\item{contentEditable}{Indicates whether the element's content is editable.}

\item{n_clicks}{An integer that represents the number of times
that this element has been clicked on.}

\item{cols}{Defines the number of columns in a textarea.}

\item{disabled}{Indicates whether the user can interact with the element.}

\item{draggable}{Defines whether the element can be dragged.}

\item{wrap}{Indicates whether the text should be wrapped.}

\item{tabIndex}{Overrides the browser's default tab order and follows the one specified instead.}

\item{n_clicks_timestamp}{An integer that represents the time (in ms since 1970)
at which n_clicks changed. This can be used to tell
which button was changed most recently.}

\item{style}{Defines CSS styles which will override styles previously set.}

\item{rows}{Defines the number of rows in a text area.}

\item{title}{Text to be displayed in a tooltip when hovering over the element.}

\item{accessKey}{Defines a keyboard shortcut to activate or add focus to the element.}

\item{id}{The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{role}{The ARIA role attribute}

\item{hidden}{Prevents rendering of given element, while keeping child elements, e.g. script elements, active.}

\item{spellCheck}{Indicates whether spell checking is allowed for the element.}

\item{form}{Indicates the form that is the owner of the element.}

\item{contextMenu}{Defines the ID of a <menu> element which will serve as the element's context menu.}

\item{minLength}{Defines the minimum number of characters allowed in the element.}

\item{readOnly}{Indicates whether the element can be edited.}

\item{key}{A unique identifier for the component, used to improve
performance by React.js while rendering components
See https://reactjs.org/docs/lists-and-keys.html for more info}

\item{maxLength}{Defines the maximum number of characters allowed in the element.}

\item{autoFocus}{The element should be automatically focused after the page loaded.}

\item{placeholder}{Provides a hint to the user of what can be entered in the field.}

\item{lang}{Defines the language used in the element.}

\item{autoComplete}{Indicates whether controls in this form can by default have their values automatically completed by the browser.}

\item{name}{Name of the element. For example used by the server to identify the fields in form submits.}

\item{required}{Indicates whether this element is required to fill out or not.}

\item{className}{Often used with CSS to style elements with common properties.}

\item{dir}{Defines the text direction. Allowed values are ltr (Left-To-Right) or rtl (Right-To-Left)}

\item{...}{wildcards of the form: `data-*` or `aria-*`}
}
