% Auto-generated: do not edit by hand
\name{htmlTrack}
\alias{htmlTrack}
\title{Track component}
\description{

}
\usage{
htmlTrack(..., contentEditable=NULL, draggable=NULL, id=NULL, children=NULL, tabIndex=NULL, n_clicks_timestamp=NULL, style=NULL, title=NULL, accessKey=NULL, label=NULL, role=NULL, srcLang=NULL, hidden=NULL, spellCheck=NULL, contextMenu=NULL, key=NULL, n_clicks=NULL, lang=NULL, src=NULL, kind=NULL, default=NULL, className=NULL, dir=NULL)
}
\arguments{
\item{contentEditable}{Indicates whether the element's content is editable.}

\item{draggable}{Defines whether the element can be dragged.}

\item{id}{The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{children}{The children of this component}

\item{tabIndex}{Overrides the browser's default tab order and follows the one specified instead.}

\item{n_clicks_timestamp}{An integer that represents the time (in ms since 1970)
at which n_clicks changed. This can be used to tell
which button was changed most recently.}

\item{style}{Defines CSS styles which will override styles previously set.}

\item{title}{Text to be displayed in a tooltip when hovering over the element.}

\item{accessKey}{Defines a keyboard shortcut to activate or add focus to the element.}

\item{label}{Specifies a user-readable title of the text track.}

\item{role}{The ARIA role attribute}

\item{srcLang}{}

\item{hidden}{Prevents rendering of given element, while keeping child elements, e.g. script elements, active.}

\item{spellCheck}{Indicates whether spell checking is allowed for the element.}

\item{contextMenu}{Defines the ID of a <menu> element which will serve as the element's context menu.}

\item{key}{A unique identifier for the component, used to improve
performance by React.js while rendering components
See https://reactjs.org/docs/lists-and-keys.html for more info}

\item{n_clicks}{An integer that represents the number of times
that this element has been clicked on.}

\item{lang}{Defines the language used in the element.}

\item{src}{The URL of the embeddable content.}

\item{kind}{Specifies the kind of text track.}

\item{default}{Indicates that the track should be enabled unless the user's preferences indicate something different.}

\item{className}{Often used with CSS to style elements with common properties.}

\item{dir}{Defines the text direction. Allowed values are ltr (Left-To-Right) or rtl (Right-To-Left)}
}
    