% Auto-generated: do not edit by hand
\docType{package}
\name{dashHtmlComponents-package}
\alias{dashHtmlComponents}
\title{Vanilla HTML Components for Dash}
\description{
Dash is a web application framework that
provides pure Python and R abstraction around HTML, CSS, and
JavaScript. Instead of writing HTML or using an HTML
templating engine, you compose your layout using R
functions within the dashHtmlComponents package. The
source for this package is on GitHub:
plotly/dash-html-components.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/plotly/dash-html-components}
  \item Report bugs at \url{https://github.com/plotly/dash-html-components/issues}
}
}
\author{
\strong{Maintainer}: Chris Parmer <chris@plot.ly>
}
